requirements_compatible.txt
python-telegram-bot==13.7
google-api-python-client==1.12.8
google-auth-oauthlib==0.4.6
gspread==3.7.0
python-dotenv==0.15.0